refactor: move RacialSkillTextHelper to client package
It resolves translations through the client Language and both call sites
are client GUIs, so it doesn't belong in common.

Also normalizes the file to tab indentation and sorts the new imports.
